Moved logging from header file to cpp file to not pollute namespace
