+++ src/app/app.cpp
@@ -1,5 +1,5 @@
 #include "app.h"
-#include "log.h"
+#include "log/log.h"
 
 #include <exception>
 #include <iostream>
